requirements.txt
# ...existing code...
Django==4.2
psycopg[binary]
djangorestframework==3.14.0
django-cors-headers==3.14.0
matplotlib==3.7.1
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
# ...existing code...